feat: checkBillsPage with p-wait-for
There is a problem with calling runInWorkerUntilTrue with a synchroneous
function. The function is called indefinitely and too fast and we miss
the worker reload sometimes, which leads to suspend of the connector
indefinitely.

I will try to reproduce this problem with template konnector

In the mean time, the good practice is to use runInWorkerUntilTrue with
p-wait-for.